[UPDATE] add sync Once inside database
