--- pkg/database/mongo.go
@@ -1,29 +1,0 @@
-package database
-
-import (
-	"context"
-	"go-crud/config"
-	"log"
-	"time"
-
-	"go.mongodb.org/mongo-driver/mongo"
-	"go.mongodb.org/mongo-driver/mongo/options"
-	"go.mongodb.org/mongo-driver/mongo/readpref"
-)
-
-func MongoConn(ctx context.Context, cfg *config.Config) *mongo.Client {
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-
-	client, err := mongo.Connect(ctx, options.Client().ApplyURI(cfg.Db.MongoDBURI))
-	if err != nil {
-		log.Fatalf("Error: Connect to database error: %s", err.Error())
-	}
-
-	if err := client.Ping(ctx, readpref.Primary()); err != nil {
-		log.Fatalf("Error: Pinging to database error: %s", err.Error())
-	}
-
-	return client
-
-}
